file renaming, fix includes
