+++ src/solver.h
@@ -1,6 +1,6 @@
 #pragma once
 
-#include "models.h"
+#include "all_models.h"
 
 #include <map>
 #include <array>
